Update modules.config
This adds the option to set more stringent quality parameters to `freebayes` and `bcftools view`.

In `freebayes`, it sets the possibility to require a minimum base quality considered for genotyping (`params.freebayes_min_basequality`) and the minimal minor allele frequency (`freebayes_minallelefreq`) that is required for freebayes to consider the variant.

In `bcftools view`, it adds the option to set two types of minimum genotype qualities. The first one is for high quality variants (`bcftools_view_high_variant_quality`): these are variants that simply pass the minimum genotype quality threshold. The second one is for medium quality variants: these are variants that pass a lower genotype quality threshold (`params.bcftools_view_medium_variant_quality`) but have a minimal number of allele observations to support the alternative genotype (`bcftools_view_minimal_allelesupport`).