Add capability of using a squash-fs image as the GTDB-Tk database.
When using singularity or apptainer, a file system image can be bind-mounted
to a container's file system.  The mount specification requires: an image,
a target path and a base path within the image file system.  This commit
allows the target path to be passed in an environment variable that can be
specified in a `containerOptions` directive.  The process interface (i.e.
inputs and outputs ) is not modified.